Add optional dependencies
